@@ -8,32 +8,26 @@ import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.DynamicCommandExceptionType;
 import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
 import com.mojang.brigadier.suggestion.SuggestionProvider;
-import dev.gegy.roles.config.PlayerRolesConfig;
 import dev.gegy.roles.Role;
 import dev.gegy.roles.api.PlayerRoleSource;
-import dev.gegy.roles.api.RoleReader;
+import dev.gegy.roles.config.PlayerRolesConfig;
 import dev.gegy.roles.override.command.CommandPermissionEvaluator;
 import dev.gegy.roles.store.PlayerRoleManager;
 import dev.gegy.roles.store.PlayerRoleSet;
 import net.minecraft.command.CommandSource;
 import net.minecraft.command.argument.GameProfileArgumentType;
-import net.minecraft.entity.Entity;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.LiteralText;
 import net.minecraft.text.MutableText;
 import net.minecraft.text.Style;
-import net.minecraft.text.Text;
 import net.minecraft.text.Texts;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.Formatting;
 
 import java.util.Collection;
-import java.util.List;
 import java.util.function.BiPredicate;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 import static net.minecraft.server.command.CommandManager.argument;
 import static net.minecraft.server.command.CommandManager.literal;
@@ -59,9 +53,9 @@ public final class RoleCommand {
                     .then(argument("targets", GameProfileArgumentType.gameProfile())
                     .then(argument("role", StringArgumentType.word()).suggests(roleSuggestions())
                     .executes(ctx -> {
-                        ServerCommandSource source = ctx.getSource();
-                        Collection<GameProfile> targets = GameProfileArgumentType.getProfileArgument(ctx, "targets");
-                        String roleName = StringArgumentType.getString(ctx, "role");
+                        var source = ctx.getSource();
+                        var targets = GameProfileArgumentType.getProfileArgument(ctx, "targets");
+                        var roleName = StringArgumentType.getString(ctx, "role");
                         return updateRoles(source, targets, roleName, PlayerRoleSet::add, "'%s' assigned to %s players");
                     })
                 )))
@@ -69,16 +63,16 @@ public final class RoleCommand {
                     .then(argument("targets", GameProfileArgumentType.gameProfile())
                     .then(argument("role", StringArgumentType.word()).suggests(roleSuggestions())
                     .executes(ctx -> {
-                        ServerCommandSource source = ctx.getSource();
-                        Collection<GameProfile> targets = GameProfileArgumentType.getProfileArgument(ctx, "targets");
-                        String roleName = StringArgumentType.getString(ctx, "role");
+                        var source = ctx.getSource();
+                        var targets = GameProfileArgumentType.getProfileArgument(ctx, "targets");
+                        var roleName = StringArgumentType.getString(ctx, "role");
                         return updateRoles(source, targets, roleName, PlayerRoleSet::remove, "'%s' removed from %s players");
                     })
                 )))
                 .then(literal("list")
                     .then(argument("target", GameProfileArgumentType.gameProfile()).executes(ctx -> {
-                        ServerCommandSource source = ctx.getSource();
-                        Collection<GameProfile> gameProfiles = GameProfileArgumentType.getProfileArgument(ctx, "target");
+                        var source = ctx.getSource();
+                        var gameProfiles = GameProfileArgumentType.getProfileArgument(ctx, "target");
                         if (gameProfiles.size() != 1) {
                             throw TOO_MANY_SELECTED.create();
                         }
@@ -91,14 +85,14 @@ public final class RoleCommand {
     // @formatter:on
 
     private static int updateRoles(ServerCommandSource source, Collection<GameProfile> players, String roleName, BiPredicate<PlayerRoleSet, Role> apply, String success) throws CommandSyntaxException {
-        Role role = getRole(roleName);
+        var role = getRole(roleName);
         assertHasPower(source, role);
 
-        PlayerRoleManager roleManager = PlayerRoleManager.get();
+        var roleManager = PlayerRoleManager.get();
         MinecraftServer server = source.getMinecraftServer();
 
         int count = 0;
-        for (GameProfile player : players) {
+        for (var player : players) {
             boolean applied = roleManager.updateRoles(server, player.getId(), roles -> apply.test(roles, role));
             if (applied) {
                 count++;
@@ -111,27 +105,27 @@ public final class RoleCommand {
     }
 
     private static int listRoles(ServerCommandSource source, GameProfile player) {
-        PlayerRoleManager roleManager = PlayerRoleManager.get();
-        MinecraftServer server = source.getMinecraftServer();
+        var roleManager = PlayerRoleManager.get();
+        var server = source.getMinecraftServer();
 
-        Collection<Role> roles = roleManager.peekRoles(server, player.getId())
+        var roles = roleManager.peekRoles(server, player.getId())
                 .stream().collect(Collectors.toList());
-        Text rolesComponent = Texts.join(roles, role -> new LiteralText(role.getName()).setStyle(Style.EMPTY.withColor(Formatting.GRAY)));
+        var rolesComponent = Texts.join(roles, role -> new LiteralText(role.getName()).setStyle(Style.EMPTY.withColor(Formatting.GRAY)));
         source.sendFeedback(new TranslatableText("Found %s roles on player: %s", roles.size(), rolesComponent), false);
 
         return Command.SINGLE_SUCCESS;
     }
 
     private static int reloadRoles(ServerCommandSource source) {
-        MinecraftServer server = source.getMinecraftServer();
+        var server = source.getMinecraftServer();
 
         server.execute(() -> {
-            List<String> errors = PlayerRolesConfig.setup();
+            var errors = PlayerRolesConfig.setup();
 
-            List<ServerPlayerEntity> players = server.getPlayerManager().getPlayerList();
-            for (ServerPlayerEntity entity : players) {
-                if (entity instanceof PlayerRoleSource) {
-                    ((PlayerRoleSource) entity).notifyPlayerRoleReload();
+            var players = server.getPlayerManager().getPlayerList();
+            for (var entity : players) {
+                if (entity instanceof PlayerRoleSource roleSource) {
+                    roleSource.notifyPlayerRoleReload();
                 }
             }
 
@@ -159,7 +153,7 @@ public final class RoleCommand {
     }
 
     private static Role getRole(String roleName) throws CommandSyntaxException {
-        Role role = PlayerRolesConfig.get().get(roleName);
+        var role = PlayerRolesConfig.get().get(roleName);
         if (role == null) throw ROLE_NOT_FOUND.create(roleName);
         return role;
     }
@@ -177,12 +171,12 @@ public final class RoleCommand {
     }
 
     private static int getHighestPowerLevel(ServerCommandSource source) {
-        Entity entity = source.getEntity();
+        var entity = source.getEntity();
         if (entity == null || CommandPermissionEvaluator.doesBypassPermissions(source)) return Integer.MAX_VALUE;
 
-        if (entity instanceof PlayerRoleSource) {
-            RoleReader roles = ((PlayerRoleSource) entity).getPlayerRoles();
-            IntStream levels = roles.stream().mapToInt(Role::getLevel);
+        if (entity instanceof PlayerRoleSource roleSource) {
+            var roles = roleSource.getPlayerRoles();
+            var levels = roles.stream().mapToInt(Role::getLevel);
             return levels.max().orElse(0);
         }
 
