@@ -3,10 +3,8 @@ package dev.gegy.roles.mixin.command;
 import com.mojang.authlib.GameProfile;
 import dev.gegy.roles.PlayerRoles;
 import dev.gegy.roles.api.PlayerRoleSource;
-import dev.gegy.roles.api.RoleReader;
 import net.minecraft.server.PlayerManager;
 import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.server.network.ServerPlayerEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
@@ -25,9 +23,9 @@ public class ServerCommandSourceMixin {
             return true;
         }
 
-        ServerPlayerEntity player = playerManager.getPlayer(profile.getId());
-        if (player instanceof PlayerRoleSource) {
-            RoleReader roles = ((PlayerRoleSource) player).getPlayerRoles();
+        var player = playerManager.getPlayer(profile.getId());
+        if (player instanceof PlayerRoleSource roleSource) {
+            var roles = roleSource.getPlayerRoles();
             return roles.test(PlayerRoles.COMMAND_FEEDBACK);
         }
 
