@@ -2,7 +2,6 @@ package dev.gegy.roles.mixin.mute;
 
 import dev.gegy.roles.PlayerRoles;
 import dev.gegy.roles.api.PlayerRoleSource;
-import dev.gegy.roles.api.RoleReader;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
 import org.spongepowered.asm.mixin.Mixin;
@@ -26,8 +25,8 @@ public class ServerPlayNetworkHandlerMixin {
             cancellable = true
     )
     private void broadcastMessage(String message, CallbackInfo ci) {
-        if (this.player instanceof PlayerRoleSource) {
-            RoleReader roles = ((PlayerRoleSource) this.player).getPlayerRoles();
+        if (this.player instanceof PlayerRoleSource roleSource) {
+            var roles = roleSource.getPlayerRoles();
             if (roles.test(PlayerRoles.MUTE)) {
                 PlayerRoles.sendMuteFeedback(this.player);
                 ci.cancel();
