@@ -1,13 +1,10 @@
 package dev.gegy.roles.override.permission;
 
 import com.mojang.serialization.Codec;
-import dev.gegy.roles.api.PermissionResult;
 import dev.gegy.roles.api.PlayerRoleSource;
 import dev.gegy.roles.api.override.RoleOverrideType;
 import me.lucko.fabric.api.permissions.v0.PermissionCheckEvent;
 import net.fabricmc.fabric.api.util.TriState;
-import net.minecraft.entity.Entity;
-import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
 
@@ -21,11 +18,10 @@ public final class PermissionKeyOverride {
     }
 
     public static void register() {
-        RoleOverrideType<PermissionKeyOverride> override = RoleOverrideType.register("permission_keys", PermissionKeyOverride.CODEC)
+        var override = RoleOverrideType.register("permission_keys", PermissionKeyOverride.CODEC)
                 .withChangeListener(owner -> {
-                    if (owner instanceof ServerPlayerEntity) {
-                        ServerPlayerEntity player = (ServerPlayerEntity) owner;
-                        MinecraftServer server = player.getServer();
+                    if (owner instanceof ServerPlayerEntity player) {
+                        var server = player.getServer();
                         if (server != null) {
                             server.getCommandManager().sendCommandTree(player);
                         }
@@ -33,10 +29,10 @@ public final class PermissionKeyOverride {
                 });
 
         PermissionCheckEvent.EVENT.register((source, permission) -> {
-            if (source instanceof ServerCommandSource) {
-                Entity entity = ((ServerCommandSource) source).getEntity();
+            if (source instanceof ServerCommandSource serverSource) {
+                var entity = serverSource.getEntity();
                 if (entity instanceof PlayerRoleSource) {
-                    PermissionResult result = ((PlayerRoleSource) entity).getPlayerRoles().test(override, permissions -> permissions.rules.test(permission));
+                    var result = ((PlayerRoleSource) entity).getPlayerRoles().test(override, permissions -> permissions.rules.test(permission));
                     return result.asTriState();
                 }
             }
