@@ -3,8 +3,6 @@ package dev.gegy.roles.mixin.mute;
 import com.mojang.brigadier.Command;
 import dev.gegy.roles.PlayerRoles;
 import dev.gegy.roles.api.PlayerRoleSource;
-import dev.gegy.roles.api.RoleReader;
-import net.minecraft.entity.Entity;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.command.TeamMsgCommand;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -18,9 +16,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class TeamMsgCommandMixin {
     @Inject(method = "execute", at = @At("HEAD"), cancellable = true)
     private static void execute(ServerCommandSource source, Text message, CallbackInfoReturnable<Integer> ci) {
-        Entity entity = source.getEntity();
+        var entity = source.getEntity();
         if (entity instanceof PlayerRoleSource && entity instanceof ServerPlayerEntity) {
-            RoleReader roles = ((PlayerRoleSource) entity).getPlayerRoles();
+            var roles = ((PlayerRoleSource) entity).getPlayerRoles();
             if (roles.test(PlayerRoles.MUTE)) {
                 PlayerRoles.sendMuteFeedback((ServerPlayerEntity) entity);
                 ci.setReturnValue(Command.SINGLE_SUCCESS);
