@@ -3,14 +3,13 @@ package dev.gegy.roles.override.command;
 import dev.gegy.roles.PlayerRoles;
 import dev.gegy.roles.api.PermissionResult;
 import dev.gegy.roles.api.RoleLookup;
-import dev.gegy.roles.api.RoleReader;
 import net.minecraft.server.command.ServerCommandSource;
 
 public final class CommandPermissionEvaluator {
     public static PermissionResult canUseCommand(ServerCommandSource source, MatchableCommand command) {
         if (doesBypassPermissions(source)) return PermissionResult.PASS;
 
-        RoleReader roles = RoleLookup.bySource(source);
+        var roles = RoleLookup.bySource(source);
         if (roles != null) {
             return roles.test(PlayerRoles.COMMANDS, m -> m.test(command));
         }
