@@ -3,9 +3,9 @@ package dev.gegy.roles.store;
 import dev.gegy.roles.api.PlayerRoleSource;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.util.NbtType;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.ListTag;
+import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtIo;
+import net.minecraft.nbt.NbtList;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.WorldSavePath;
@@ -18,7 +18,6 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.nio.file.Path;
 import java.util.Objects;
 import java.util.UUID;
 import java.util.function.Function;
@@ -40,7 +39,7 @@ public final class PlayerRoleManager {
         });
 
         ServerLifecycleEvents.SERVER_STOPPED.register(server -> {
-            PlayerRoleManager instance = PlayerRoleManager.instance;
+            var instance = PlayerRoleManager.instance;
             if (instance != null) {
                 instance.close(server);
             }
@@ -50,8 +49,8 @@ public final class PlayerRoleManager {
 
     private static PlayerRoleManager open(MinecraftServer server) {
         try {
-            Path path = server.getSavePath(WorldSavePath.PLAYERDATA).resolve("player_roles");
-            PlayerIndexedDatabase database = PlayerIndexedDatabase.open(path);
+            var path = server.getSavePath(WorldSavePath.PLAYERDATA).resolve("player_roles");
+            var database = PlayerIndexedDatabase.open(path);
             return new PlayerRoleManager(database);
         } catch (IOException e) {
             throw new RuntimeException("failed to open player roles database");
@@ -63,13 +62,13 @@ public final class PlayerRoleManager {
     }
 
     public void onPlayerJoin(ServerPlayerEntity player) {
-        PlayerRoleSource roleOwner = (PlayerRoleSource) player;
+        var roleOwner = (PlayerRoleSource) player;
         this.loadRolesInto(player.getUuid(), roleOwner.getPlayerRoles());
     }
 
     public void onPlayerLeave(ServerPlayerEntity player) {
-        PlayerRoleSource roleOwner = (PlayerRoleSource) player;
-        PlayerRoleSet roles = roleOwner.getPlayerRoles();
+        var roleOwner = (PlayerRoleSource) player;
+        var roles = roleOwner.getPlayerRoles();
         if (roles.isDirty()) {
             try {
                 this.saveRoles(player.getUuid(), roles);
@@ -82,7 +81,7 @@ public final class PlayerRoleManager {
 
     private void close(MinecraftServer server) {
         try {
-            for (ServerPlayerEntity player : server.getPlayerManager().getPlayerList()) {
+            for (var player : server.getPlayerManager().getPlayerList()) {
                 this.onPlayerLeave(player);
             }
         } finally {
@@ -92,7 +91,7 @@ public final class PlayerRoleManager {
 
     private void loadRolesInto(UUID uuid, PlayerRoleSet roles) {
         try {
-            ByteBuffer bytes = this.database.get(uuid);
+            var bytes = this.database.get(uuid);
             if (bytes != null) {
                 try {
                     deserializeRoles(roles, bytes);
@@ -108,7 +107,7 @@ public final class PlayerRoleManager {
 
     private void saveRoles(UUID uuid, PlayerRoleSet roles) throws IOException {
         if (!roles.isEmpty()) {
-            ByteBuffer bytes = serializeRoles(roles);
+            var bytes = serializeRoles(roles);
             this.database.put(uuid, bytes);
         } else {
             this.database.remove(uuid);
@@ -116,34 +115,34 @@ public final class PlayerRoleManager {
     }
 
     private static ByteBuffer serializeRoles(PlayerRoleSet roles) throws IOException {
-        CompoundTag nbt = new CompoundTag();
+        var nbt = new NbtCompound();
         nbt.put("roles", roles.serialize());
 
-        try (ByteArrayOutputStream output = new ByteArrayOutputStream()) {
+        try (var output = new ByteArrayOutputStream()) {
             NbtIo.writeCompressed(nbt, output);
             return ByteBuffer.wrap(output.toByteArray());
         }
     }
 
     private static void deserializeRoles(PlayerRoleSet roles, ByteBuffer bytes) throws IOException {
-        try (ByteArrayInputStream input = new ByteArrayInputStream(bytes.array())) {
-            CompoundTag nbt = NbtIo.readCompressed(input);
+        try (var input = new ByteArrayInputStream(bytes.array())) {
+            var nbt = NbtIo.readCompressed(input);
             roles.deserialize(nbt.getList("roles", NbtType.STRING));
         }
     }
 
-    public void addLegacyRoles(PlayerRoleSource owner, ListTag nbt) {
-        PlayerRoleSet roles = owner.getPlayerRoles();
+    public void addLegacyRoles(PlayerRoleSource owner, NbtList nbt) {
+        var roles = owner.getPlayerRoles();
         roles.deserialize(nbt);
         roles.setDirty(true);
     }
 
     public <R> R updateRoles(MinecraftServer server, UUID uuid, Function<PlayerRoleSet, R> update) {
-        PlayerRoleSource roleOwner = getRoleOwner(server, uuid);
+        var roleOwner = getRoleOwner(server, uuid);
         if (roleOwner != null) {
             return update.apply(roleOwner.getPlayerRoles());
         } else {
-            PlayerRoleSet roles = new PlayerRoleSet(null);
+            var roles = new PlayerRoleSet(null);
             this.loadRolesInto(uuid, roles);
 
             try {
@@ -161,11 +160,11 @@ public final class PlayerRoleManager {
     }
 
     public PlayerRoleSet peekRoles(MinecraftServer server, UUID uuid) {
-        PlayerRoleSource roleOwner = getRoleOwner(server, uuid);
+        var roleOwner = getRoleOwner(server, uuid);
         if (roleOwner != null) {
             return roleOwner.getPlayerRoles();
         } else {
-            PlayerRoleSet roles = new PlayerRoleSet(null);
+            var roles = new PlayerRoleSet(null);
             this.loadRolesInto(uuid, roles);
             return roles;
         }
@@ -173,7 +172,7 @@ public final class PlayerRoleManager {
 
     @Nullable
     private static PlayerRoleSource getRoleOwner(MinecraftServer server, UUID uuid) {
-        ServerPlayerEntity player = server.getPlayerManager().getPlayer(uuid);
+        var player = server.getPlayerManager().getPlayer(uuid);
         return (PlayerRoleSource) player;
     }
 }
